% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{plot a map of accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a number code for a state}

\item{year}{a list of years}
}
\value{
a data frame with year and count
}
\description{
plot a map of accidents for the given state and year
}
\note{
warns when the year is not found in the fars data

stops when the state is not found in the fars data

warns when no accidents found
}
\examples{
\dontrun{d <- fars_map_states(state.num = 9, years = "2015") }
}
