% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{read a fars file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the path to the file}
}
\value{
a data frame
}
\description{
This funtion reads the file located at \code{filename}.
}
\note{
stops if \code{filename} does not exist
}
\examples{
\dontrun{data <- fars_read(make_filename("2013"))}
}
