% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{months and years in the fars data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a list of years}
}
\value{
a data frame with the month and year
}
\description{
read and print the months and years from fars data for a given list of years
}
\note{
warns when a year is not found in the fars data
}
\examples{
\dontrun{d <- fars_read_years(years = c("2015","2016")) }
}
