% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{read and summarize fars data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a list of years}
}
\value{
a data frame with year and count
}
\description{
read fars data and summize annual data for a given list of years
}
\note{
warns when a year is not found in the fars data
}
\examples{
\dontrun{d <- fars_summarize_years(years = c("2015","2016")) }
}
