% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{generate a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a string or integer. It is converted to an integer.}
}
\value{
string of the generaed filename
}
\description{
generate a string in the form of 'accident_<year>.csv.bz2'
}
\examples{
make_filename("2017")
}
